# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/agent/pros/{id}": {
      "get": {
        "operationId": "AgentController_getProById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Agent ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Professional profile details"
          }
        },
        "summary": "Get single professional agent details for public profile",
        "tags": [
          "Agent"
        ]
      }
    },
...
    "/agent/{id}": {
      "get": {
        "operationId": "AgentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Agent ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent details"
          }
        },
        "summary": "Get agent by ID",
        "tags": [
          "Agent"
        ]
      },
      "patch": {
        "operationId": "AgentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Agent ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent updated"
          }
        },
        "summary": "Update agent details",
        "tags": [
          "Agent"
        ]
      },
      "delete": {
        "operationId": "AgentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Agent ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent removed"
          }
        },
        "summary": "Delete agent (soft delete)",
        "tags": [
          "Agent"
        ]
      }
    },
...
    "/marketproduct": {
      "post": {
        "description": "Uploads product metadata alongside architecture, structural, electrical, and mechanical plan files",
        "operationId": "MarketproductController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Modern 4-Bedroom Duplex"
                  },
                  "description": {
                    "type": "string",
                    "example": "Contemporary architectural design with complete structural calculations."
                  },
                  "price": {
                    "type": "number",
                    "example": 150000
                  },
                  "category": {
                    "type": "string",
                    "example": "Residential"
                  },
                  "subCategory": {
                    "type": "string",
                    "example": "Duplex"
                  },
                  "productImage": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Product preview photos / 3D renders"
                  },
                  "architecturalPlan": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Architectural blueprints / CAD / PDF files"
                  },
                  "structuralPlan": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Structural engineering drawings"
                  },
                  "electricalPlan": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Electrical layout plans"
                  },
                  "mechanicalPlan": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Mechanical / plumbing plans"
                  }
                },
                "required": [
                  "title",
                  "price"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Marketplace product listed successfully"
          },
          "401": {
            "description": "Unauthorized - Agent token required"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new marketplace design product listing (Agent only)",
        "tags": [
          "Marketplace Product"
        ]
      },
      "get": {
        "description": "Retrieve marketplace products with filters for category, price range, search keywords, and pagination",
        "operationId": "MarketproductController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "planType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Residential",
                "Commercial",
                "Industrial"
              ]
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending_review",
                "approved",
                "rejected"
              ]
            }
          },
          {
            "name": "buildingGuides",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "numBedrooms",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "numBathrooms",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "numFloors",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "area",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "designStyle",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filtered marketplace product listing"
          }
        },
        "summary": "Search and filter marketplace products",
        "tags": [
          "Marketplace Product"
        ]
      }
    },
    "/marketproduct/agent/{agentId}": {
      "get": {
        "operationId": "MarketproductController_findAllByAgent",
        "parameters": [
          {
            "name": "agentId",
            "required": true,
            "in": "path",
            "description": "Agent ID UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "planType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Residential",
                "Commercial",
                "Industrial"
              ]
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending_review",
                "approved",
                "rejected"
              ]
            }
          },
          {
            "name": "buildingGuides",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "numBedrooms",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "numBathrooms",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "numFloors",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "area",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "designStyle",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent product listing"
          }
        },
        "summary": "Get all marketplace products published by an agent",
        "tags": [
          "Marketplace Product"
        ]
      }
    },
    "/marketproduct/{id}": {
      "get": {
        "operationId": "MarketproductController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product details"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "summary": "Get marketplace product details by ID",
        "tags": [
          "Marketplace Product"
        ]
      },
      "patch": {
        "operationId": "MarketproductController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMarketproductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update marketplace product (Agent owner only)",
        "tags": [
          "Marketplace Product"
        ]
      },
      "delete": {
        "operationId": "MarketproductController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete marketplace product (Agent owner only)",
        "tags": [
          "Marketplace Product"
        ]
      }
    },
    "/marketproduct/{id}/rate": {
      "post": {
        "operationId": "MarketproductController_rate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review submitted"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Rate and review a marketplace product",
        "tags": [
          "Marketplace Product"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mejarc API",
    "description": "Comprehensive REST API documentation for Mejarc Backend services, including Authentication, User Management, Agent Management, Marketplace, Custom Design Wizard, Orders, Payments, Wallet & Withdrawals, Notifications, and Real-Time Chat.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter your Bearer token in format: Bearer <token>",
        "in": "header"
      }
    },
    "schemas": {
...
      "UpdateAgentDto": {
        "type": "object",
        "properties": {
          "businessName": {
            "type": "string"
          },
          "yearsOfExperience": {
            "type": "number",
            "minimum": 0,
            "maximum": 70
          },
          "preferredTitle": {
            "type": "string"
          },
          "specialization": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "portfolioLink": {
            "type": "string"
          },
          "profilePicture": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "idType": {
            "type": "string"
          },
          "idNumber": {
            "type": "string"
          },
          "idDocument": {
            "type": "string"
          },
          "architectCert": {
            "type": "string"
          },
          "bankName": {
            "type": "string"
          },
          "accountNumber": {
            "type": "string"
          },
          "accountHolderName": {
            "type": "string"
          },
          "rejectionReason": {
            "type": "string"
          }
        }
      },
...
      "UpdateMarketproductDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "category": {
            "enum": [
              "Building Plan",
              "Interior Design",
              "Landscape Design"
            ],
            "type": "string"
          },
          "planType": {
            "enum": [
              "Residential",
              "Commercial",
              "Industrial"
            ],
            "type": "string"
          },
          "numBedrooms": {
            "type": "string"
          },
          "numBathrooms": {
            "type": "string"
          },
          "numFloors": {
            "type": "string"
          },
          "area": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "fileType": {
            "enum": [
              "PDF only",
              "CAD + PDF",
              "3D Model + CAD + PDF"
            ],
            "type": "string"
          },
          "drawingSet": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "addOns": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "productImage": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "architecturalPlan": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "structuralPlan": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "electricalPlan": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mechanicalPlan": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "architecturalPlanFileType": {
            "type": "string"
          },
          "structuralPlanFileType": {
            "type": "string"
          },
          "electricalPlanFileType": {
            "type": "string"
          },
          "mechanicalPlanFileType": {
            "type": "string"
          }
        }
      },
      "RateProductDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "rating"
        ]
      },
...
    }
  }
}